Rename EntityPersister to PostgresDriver.
